requirement.txt
Pyarrow
pandas
matplotlib
openpyxl
tiktoken
openai
scikit-learn
tqdm
pysocks
fitz
wordcloud
paramiko
PyPDF2
PyYAML